# backlog/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IndexCreate": {
        "additionalProperties": false,
        "properties": {
          "column_keys": {
            "items": {
              "type": "string"
            },
            "title": "Column Keys",
            "type": "array"
          },
          "index_kind": {
            "default": "btree",
            "title": "Index Kind",
            "type": "string"
          },
          "is_unique": {
            "default": false,
            "title": "Is Unique",
            "type": "boolean"
          }
        },
        "required": [
          "column_keys"
        ],
        "title": "IndexCreate",
        "type": "object"
      },
...
      "RowCreate": {
        "additionalProperties": false,
        "properties": {
          "values": {
            "additionalProperties": true,
            "title": "Values",
            "type": "object"
          }
        },
        "title": "RowCreate",
        "type": "object"
      },
      "RowOut": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "id",
          "payload"
        ],
        "title": "RowOut",
        "type": "object"
      },
      "RowPatch": {
        "additionalProperties": false,
        "properties": {
          "values": {
            "additionalProperties": true,
            "title": "Values",
            "type": "object"
          }
        },
        "title": "RowPatch",
        "type": "object"
      },
...
      "TableCreate": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "natural_key": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Natural Key"
          },
          "ownership": {
            "default": "owned",
            "enum": [
              "owned",
              "bound"
            ],
            "title": "Ownership",
            "type": "string"
          },
          "projection_predicate": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection Predicate"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "label"
        ],
        "title": "TableCreate",
        "type": "object"
      },
      "TableDescribe": {
        "additionalProperties": false,
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnOut"
            },
            "title": "Columns",
            "type": "array"
          },
          "indexes": {
            "items": {
              "$ref": "#/components/schemas/IndexOut"
            },
            "title": "Indexes",
            "type": "array"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/RelationOut"
            },
            "title": "Relations",
            "type": "array"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionOut"
            },
            "title": "Sections",
            "type": "array"
          },
          "table": {
            "$ref": "#/components/schemas/TableOut"
          }
        },
        "required": [
          "table",
          "columns",
          "sections",
          "relations",
          "indexes"
        ],
        "title": "TableDescribe",
        "type": "object"
      },
      "TableOut": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "namespace": {
            "title": "Namespace",
            "type": "string"
          },
          "natural_key": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Natural Key"
          },
          "ownership": {
            "title": "Ownership",
            "type": "string"
          },
          "projection_predicate": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection Predicate"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "namespace",
          "slug",
          "label",
          "ownership"
        ],
        "title": "TableOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Self-describing dynamic virtual database \u2014 the agent's persistent state space.",
    "title": "ForkTex Grid",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/grid/rows/{row_id}": {
      "delete": {
        "operationId": "archive_row_grid_rows__row_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "row_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Row Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Row",
        "tags": [
          "grid"
        ]
      },
      "get": {
        "operationId": "get_row_grid_rows__row_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "row_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Row Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Row",
        "tags": [
          "grid"
        ]
      },
      "patch": {
        "operationId": "patch_row_grid_rows__row_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "row_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Row Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RowPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Row",
        "tags": [
          "grid"
        ]
      }
    },
    "/grid/tables": {
      "get": {
        "operationId": "list_tables_grid_tables_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-grid-namespace",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Grid-Namespace",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TableOut"
                  },
                  "title": "Response List Tables Grid Tables Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tables",
        "tags": [
          "grid"
        ]
      },
      "post": {
        "operationId": "create_table_grid_tables_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-grid-namespace",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Grid-Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TableCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Table",
        "tags": [
          "grid"
        ]
      }
    },
    "/grid/tables/{slug}": {
      "get": {
        "operationId": "describe_table_grid_tables__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-grid-namespace",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Grid-Namespace",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableDescribe"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Describe Table",
        "tags": [
          "grid"
        ]
      }
    },
    "/grid/tables/{slug}/columns": {
      "post": {
        "operationId": "add_column_grid_tables__slug__columns_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-grid-namespace",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Grid-Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ColumnCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColumnOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Column",
        "tags": [
          "grid"
        ]
      }
    },
    "/grid/tables/{slug}/indexes": {
      "post": {
        "operationId": "add_index_grid_tables__slug__indexes_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-grid-namespace",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Grid-Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Index",
        "tags": [
          "grid"
        ]
      }
    },
    "/grid/tables/{slug}/query": {
      "post": {
        "operationId": "query_grid_tables__slug__query_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-grid-namespace",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Grid-Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/QueryRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Query",
        "tags": [
          "grid"
        ]
      }
    },
    "/grid/tables/{slug}/reconcile": {
      "post": {
        "operationId": "reconcile_grid_tables__slug__reconcile_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-grid-namespace",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Grid-Namespace",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "title": "Response Reconcile Grid Tables  Slug  Reconcile Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reconcile",
        "tags": [
          "grid"
        ]
      }
    },
    "/grid/tables/{slug}/relations": {
      "post": {
        "operationId": "add_relation_grid_tables__slug__relations_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-grid-namespace",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Grid-Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Relation",
        "tags": [
          "grid"
        ]
      }
    },
    "/grid/tables/{slug}/rows": {
      "post": {
        "operationId": "create_row_grid_tables__slug__rows_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-grid-namespace",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Grid-Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RowCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RowOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Row",
        "tags": [
          "grid"
        ]
      }
    },
    "/grid/tables/{slug}/rows/{row_id}/links": {
      "get": {
        "operationId": "list_links_grid_tables__slug__rows__row_id__links_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "row_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Row Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "relation_key",
            "required": true,
            "schema": {
              "title": "Relation Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-grid-namespace",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Grid-Namespace",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RowOut"
                  },
                  "title": "Response List Links Grid Tables  Slug  Rows  Row Id  Links Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Links",
        "tags": [
          "grid"
        ]
      }
    },
    "/grid/tables/{slug}/rows/{row_id}/relate": {
      "post": {
        "operationId": "relate_grid_tables__slug__rows__row_id__relate_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "row_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Row Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-grid-namespace",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Grid-Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Relate",
        "tags": [
          "grid"
        ]
      }
    },
    "/grid/tables/{slug}/sections": {
      "post": {
        "operationId": "add_section_grid_tables__slug__sections_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-grid-namespace",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Grid-Namespace",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Section",
        "tags": [
          "grid"
        ]
      }
    },
...
  }
}